requirements.txt
streamlit==1.32.0
google-cloud-secret-manager==2.18.3
google-generativeai==0.4.1
google-cloud-firestore==2.14.0
google-cloud-logging==3.9.0
google-cloud-storage==2.14.0
google-cloud-monitoring==2.19.0
pytest==8.1.1
pytest-asyncio==0.23.5
python-dotenv==1.0.1
